Create tidy FIA tables

This function queries the FIA database, by state abbreviation(s) or area of interest,and returns a list of tidy data objects including the TREE, PLOT, COND, and SURVEY tables.

tidy_fia(
  states = NULL,
  aoi = NULL,
  postgis = TRUE,
  table_names = c("plot", "subplot", "cond", "tree", "survey")
)

Arguments

states
a character vector of state abbreviations, ignored if aoi is supplied.

aoi
sf object containing area of interest

postgis
logical if true, query PostGIS database instead of downloading csvs from FIA Datamart.

table_names
list of FIA tables to download. Tables must be identified by their 'Oracle Table Name' as described in the 'Index of Tables' in FIADB User Guide.

Value

a list object containing tidy data